+++ src/main/java/com/main/controller/order/remove/OrderPaidRemoveStrategy.java
@@ -2,6 +2,7 @@ package com.main.controller.order.remove;
 
 import com.main.entities.account.BalanceEntity;
 import com.main.entities.order.OrderEntity;
+import com.main.entities.order.OrderType;
 import com.main.services.AccountService;
 import com.main.services.FileService;
 import com.main.services.OrderService;
@@ -21,9 +22,11 @@ public class OrderPaidRemoveStrategy implements OrderRemoveStrategy {
 
     @Override
     public boolean remove(OrderEntity orderEntity, BalanceEntity balanceEntity) {
-        final boolean isAttachedFilesRemoved = fileService.removeFilesByOrderId(orderEntity.getOrderId());
-        if (!isAttachedFilesRemoved) {
-            return false;
+        if (orderEntity.getOrderType() == OrderType.PRINT) {
+            final boolean isAttachedFilesRemoved = fileService.removeFilesByOrderId(orderEntity.getOrderId());
+            if (!isAttachedFilesRemoved) {
+                return false;
+            }
         }
         final BigDecimal newAccountBalance = balanceEntity.getAccountBalance().add(
                 orderEntity.getOrderAmount()
